Changes to dbus spec.
As we will be representing the types in the keyspace xml, it makes zero sense
to have types on the wire, so this commit removes TypedVariant. all key
values are now sent in a plain map of strings to variants. To cope with
keys who's values are currently undeterminable, we've added an extra
argument to the return from Get and in the Changed signals with an array
of keys whos values are currently undeterminable.