Added simple enum support, removed unnecessary (constant) path attribute.
Note that implementation of the ENUM type is not perfect as apparently W3C XML Schema
does not support conditional element content depending on an attribute value. RELAX-NG
could be used for such approach.
In this simple approach the enumerator is an optional element of <key>. Therefore the application
should check that a <key type=ENUM> actually has enumerator elements. For the other types: INT, DOUBLE,
etc. the application should simply ignore the possibly present enumerator elements.