libcontextsubscriber: try the new protocol only if needed.
If the old protocol fails because of some generic error, for example
not enough memory, no dbus, etc. then there is no point in trying the
new one.  The only case when the trying the new one makes sense is
when the call failed with QDBusError::UnknownObject (no manager object
found).

Signed-off-by: Marja Hassinen <ext-marja.2.hassinen@nokia.com>