Fixing mess created by waitForSubscription(bool block).
The problem was related to the "with context commander, try out both
the old and the new protocol" feature. It caused a crash in
subscribe (subscriberInterface == 0). We had forgotten that the old protocol
doesn't work, because blockUntilReady() has called reset().

A problem: programs using waitForSubscription(true) don't work with the
graphical Context Commander which uses the old protocol (blockUntilSubscribed
is not implemented properly).